feat(home): add notification badge and screen
- Created notification badge widget
- Implemented notification screen
- Added notification reading functionality
- Integrated with shared preferences
- Improved UI/UX for notifications
